+++ packages/components-angular/src/app/examples/index.ts
@@ -1,9 +1,16 @@
+import { AccordionExampleComponent } from './accordion-example.component';
 import { TableExampleAdvancedComponent } from './table-example-advanced.component';
 import { TableExampleBasicComponent } from './table-example-basic.component';
 import { TableExampleSortingComponent } from './table-example-sorting.component';
 
-export const examples = [TableExampleAdvancedComponent, TableExampleBasicComponent, TableExampleSortingComponent];
+export const examples = [
+  AccordionExampleComponent,
+  TableExampleAdvancedComponent,
+  TableExampleBasicComponent,
+  TableExampleSortingComponent,
+];
 
+export * from './accordion-example.component';
 export * from './table-example-advanced.component';
 export * from './table-example-basic.component';
 export * from './table-example-sorting.component';