+++ packages/components-react/projects/jsdom-polyfill/src/index.js
@@ -1,2 +1,5 @@
+import { ResizeObserver } from '@juggle/resize-observer';
+global.ResizeObserver = ResizeObserver;
+
 // workaround for WebComponentManager not working in JSDOM
 require('./lib/loader.cjs').defineCustomElements();