+++ packages/components-react/tests/e2e/specs/examples.e2e.ts
@@ -1,50 +1,28 @@
-import { ConsoleMessage, Page } from 'puppeteer';
-import { goto } from '../helpers';
+import { Page } from 'puppeteer';
+import { goto, initConsoleObserver, getConsoleErrorsAmount } from '../helpers';
 import { browser } from '../config';
 
 describe('examples', () => {
   let page: Page;
-  let consoleMessages: ConsoleMessage[] = [];
 
   beforeEach(async () => {
     page = await browser.newPage();
-
-    consoleMessages = []; // reset
-    page.on('console', (msg) => {
-      consoleMessages.push(msg);
-      if (msg.type() === 'error') {
-        const { description } = msg.args()[0]['_remoteObject'];
-        if (description) {
-          console.log(description);
-        }
-      }
-    });
+    initConsoleObserver(page);
   });
   afterEach(async () => await page.close());
 
-  const getErrorsAmount = () => consoleMessages.filter((x) => x.type() === 'error').length;
+  for (const example of [
+    'accordion-example',
+    'table-example-basic',
+    'table-example-sorting',
+    'table-example-advanced',
+  ]) {
+    it(`should work without error for ${example}`, async () => {
+      await goto(page, example);
+      expect(getConsoleErrorsAmount()).toBe(0);
 
-  it('table example basic should work without errors', async () => {
-    await goto(page, 'table-example-basic');
-    expect(getErrorsAmount()).toBe(0);
-
-    await page.evaluate(() => console.error('test error'));
-    expect(getErrorsAmount()).toBe(1);
-  });
-
-  it('table example sorting should work without errors', async () => {
-    await goto(page, 'table-example-sorting');
-    expect(getErrorsAmount()).toBe(0);
-
-    await page.evaluate(() => console.error('test error'));
-    expect(getErrorsAmount()).toBe(1);
-  });
-
-  it('table example advanced should work without errors', async () => {
-    await goto(page, 'table-example-advanced');
-    expect(getErrorsAmount()).toBe(0);
-
-    await page.evaluate(() => console.error('test error'));
-    expect(getErrorsAmount()).toBe(1);
-  });
+      await page.evaluate(() => console.error('test error'));
+      expect(getConsoleErrorsAmount()).toBe(1);
+    });
+  }
 });