+++ packages/components/src/components/layout/table/table-head-cell/table-head-cell.tsx
@@ -30,7 +30,7 @@ export class TableHeadCell {
     return (
       <Host scope="col" role="columnheader" aria-sort={getAriaSort(this.sort)}>
         {active !== undefined && direction !== undefined ? (
-          <button onClick={this.handleButtonClick}>
+          <button onClick={this.onButtonClick}>
             <slot />
             <PrefixedTagNames.pIcon
               class={{
@@ -51,7 +51,7 @@ export class TableHeadCell {
     );
   }
 
-  private handleButtonClick = (): void => {
+  private onButtonClick = (): void => {
     this.host.dispatchEvent(
       new CustomEvent<SortingChangeEvent>(SORT_EVENT_NAME, createSortedEventInitDictDetail(this.sort))
     );