+++ packages/components/src/components/layout/table/table/table.tsx
@@ -82,7 +82,7 @@ export class Table {
                 hide-label="true"
                 size="inherit"
                 icon="arrow-head-right"
-                onClick={this.handleScrollClick}
+                onClick={this.onScrollClick}
               >
                 Next
               </PrefixedTagNames.pButtonPure>
@@ -121,7 +121,7 @@ export class Table {
     this.intersectionObserver.observe(this.tableElement); // to check if table is scrollable in general
   };
 
-  private handleScrollClick = (): void => {
+  private onScrollClick = (): void => {
     scrollElementBy(this.scrollAreaElement, getScrollByX(this.scrollAreaElement));
   };
 }