+++ packages/components/CHANGELOG.md
@@ -9,13 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Unreleased]
 
 #### Added
+- `Checkbox Wrapper`, `Radio Button Wrapper`, `Select Wrapper`, `Textarea Wrapper` and `Text Field Wrapper` now reflect changes of the `required` attribute on their child component
 - `multiline` property to `Table Cell`
-
-#### Changed
-- Removed initial delay of `Banner`
 
 #### Fixed
 - `Table Head Cell` uses semi bold instead of bold as font weight
+
+### [2.3.0-beta.1] - 2021-07-08
+
+#### Added
+- `Accordion`
+
+#### Changed
+- Removed initial delay of `Banner`
 
 ### [2.3.0-beta.0] - 2021-07-01